+++ example/counter_notifier_example.dart
@@ -1,0 +1,9 @@
+import 'package:syncx/syncx.dart';
+
+// Example Notifier: simple counter
+class CounterNotifier extends Notifier<int> {
+  CounterNotifier() : super(0);
+
+  void increment() => setState(state + 1);
+  void decrement() => setState(state - 1);
+}